+++ slack-bot-server/commands.rb
@@ -1,2 +1,2 @@
 require 'slack-bot-server/commands/help'
-require 'slack-bot-server/commands/children'
+require 'slack-bot-server/commands/kids'