+++ slack-bot-server/models.rb
@@ -1,3 +1,3 @@
 require 'slack-bot-server/models/team'
-require 'slack-bot-server/models/missing_child'
-require 'slack-bot-server/models/missing_children_notifier'
+require 'slack-bot-server/models/missing_kid'
+require 'slack-bot-server/models/missing_kids_notifier'