+++ spec/api/endpoints/root_endpoint_spec.rb
@@ -7,7 +7,7 @@ describe Api::Endpoints::RootEndpoint do
     get '/api'
     expect(last_response.status).to eq 200
     links = JSON.parse(last_response.body)['_links']
-    expect(links.keys.sort).to eq(%w(self status team teams missing_child missing_children).sort)
+    expect(links.keys.sort).to eq(%w(self status team teams missing_kid missing_kids).sort)
   end
   it 'follows all links' do
     get '/api'