+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Changelog
 
-* 2016/4/24: Respond to `@missing children` - [@dblock](https://github.com/dblock).
+* 2016/4/24: Respond to `@missing kids` - [@dblock](https://github.com/dblock).
 * 2016/4/24: Fix [#1](https://github.com/dblock/slack-amber-alert/issues/1): don't notify a second time after joining channel - [@dblock](https://github.com/dblock).
 * 2016/4/18: Initial public release - [@dblock](https://github.com/dblock).